src/auth/auth.guard.ts
import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
import { Request } from 'express';
import { JwtService } from '@nestjs/jwt';
import { ConfigService } from '@nestjs/config';

@Injectable()
export class AuthGuard implements CanActivate {
	constructor(
		private readonly configService: ConfigService,
		private readonly jwtService: JwtService
	) {}

	async canActivate(context: ExecutionContext): Promise<boolean> {
		const request = context.switchToHttp().getRequest();
		const token = this.extractTokenFromHeader(request);

		if (!token) throw new UnauthorizedException();

		try {
			const payload = await this.jwtService.verifyAsync(token, { secret: this.configService.get('JWT_CONSTANTS') });

			request['user'] = payload;
		} catch (e) {
			throw new UnauthorizedException();
		}

		return true;
	}

	extractTokenFromHeader(request: Request): string | undefined {
		const [type, token] = request.headers.authorization.split(' ') ?? [];
		return type === 'Bearer' ? token : undefined;
	}
}
